refactor(core-components-select): split select, move types (#137)
* refactor(core-components-select): split select, move types

* fix(core-components-select): fix highlight on typing feature

* fix(core-components-select): pr fixes
