+++ packages/select/package.json
@@ -5,6 +5,7 @@
     "keywords": [],
     "license": "ISC",
     "main": "dist/index.js",
+    "types": "dist/typings.d.ts",
     "files": [
         "dist"
     ],
