+++ packages/select/src/components/arrow/index.ts
@@ -1,0 +1,1 @@
+export * from './Component';
