+++ packages/select/src/components/base-select/index.module.css
@@ -1,0 +1,40 @@
+@import '../../../../themes/src/default.css';
+@import '../../vars.css';
+
+.component {
+    display: inline-block;
+    min-width: 240px;
+    position: relative;
+}
+
+.fieldWrapper {
+    position: relative;
+}
+
+.popover {
+    min-width: 100%;
+    border: none;
+    box-shadow: var(--select-popover-box-shadow);
+    border-radius: var(--select-popover-border-radius);
+    overflow: hidden;
+}
+
+.listWrapper {
+    width: 100%;
+    outline: none;
+}
+
+.nativeSelect {
+    opacity: 0;
+    overflow: hidden;
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    max-width: 100%;
+    min-width: 0;
+    height: 100%;
+    cursor: pointer;
+    z-index: 1;
+}
