+++ packages/select/src/components/base-select/index.ts
@@ -1,0 +1,1 @@
+export * from './Component';
